% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{condDataS1}
\alias{condDataS1}
\title{Formatted conductivity time-series data.}
\format{
A list with 2663 rows and 1 variable:
\describe{
  \item{x}{specific conductance of water, specificConductance}
  ...
}
}
\source{
\url{http://data.neoninc.org/home/}
}
\usage{
condDataS1
}
\description{
A list containing a vector of conductivity data
}
\keyword{datasets}
