% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constTime.R
\name{constTime}
\alias{constTime}
\title{Constant rate injection half plateau specific conductance location}
\usage{
constTime(reaCond, plotName, bPlot)
}
\arguments{
\item{reaCond}{Specific conductance values of the stream during the experiment [uS/cm @ 25 C]}

\item{plotName}{Plot title to be displayed [string]}

\item{bPlot}{User input to display a visual check, default is FALSE [boolean]}
}
\value{
This function returns the peak tracer location [integer]
}
\description{
This function determines the location of the half plateau specific conductance for use 
calculating travel time between an upstream and downstream sensor set.
}
\examples{
#Using an example file
s1peakLoc <- constTime(reaCond = condDataS1, plotName = "Example_S1", bPlot = TRUE)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
def.calc.travelTime for calculating travel times
}
\author{
Kaelin M. Cawley \email{kcawley@battelleecology.org} \cr
}
\keyword{metabolism}
\keyword{reaeration,}
\keyword{rivers,}
\keyword{streams,}
\keyword{surface}
\keyword{time,}
\keyword{travel}
\keyword{velocity,}
\keyword{water,}
