% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.calc.peakTime.R
\name{def.calc.peakTime}
\alias{def.calc.peakTime}
\title{Constant rate or slug travel time between two stations}
\usage{
def.calc.peakTime(loggerData, currEventID, injectionType)
}
\arguments{
\item{loggerData}{User input of the R data object holding the conductivity time series
for a site, date, and station [string]}

\item{currEventID}{User input of the eventID of the tracer experiment [string]}

\item{injectionType}{User input of the injection type either "constant" or "slug" [string]}
}
\value{
This function returns the peak tracer timestamp [dateTime]
}
\description{
This function determines the timestamp of the peak tracer conductance for use
calculating travel time between an upstream and downstream sensor set.
}
\examples{
#Using an example file
#travelTime <- def.calc.travelTime(
#dataDir = paste(path.package("reaRate"),"inst\\extdata", sep = "\\"),
#currEventID = "GUIL.20150129", injectionType = "constant", bPlot = T)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
def.calc.reaeration.R for calculating reaeration rates
}
\author{
Kaelin M. Cawley \email{kcawley@battelleecology.org} \cr
}
\keyword{metabolism}
\keyword{reaeration,}
\keyword{rivers,}
\keyword{streams,}
\keyword{surface}
\keyword{time,}
\keyword{travel}
\keyword{velocity,}
\keyword{water,}
