% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.calc.reaeration.R
\name{def.calc.reaeration}
\alias{def.calc.reaeration}
\title{Reaeration rate and Schmidt number calculations}
\usage{
def.calc.reaeration(
  inputFile = NULL,
  lossRateSF6 = "lossRateSF6",
  peakMaxVelocity = "peakMaxVelocity",
  meanDepth = "meanDepth",
  meanTemp = "meanTemp",
  outputSuffix = NULL
)
}
\arguments{
\item{inputFile}{Name of the data fram containing the information needed to calculate the
reaeration parameters. If the headers are named: "injectionType", "eventID",
"stationToInjectionDistance", "plateauGasConc", "corrPlatSaltConc", "hoboSampleID",
"wettedWidth", respectively, no other inputs are required. Otherwise, the names of the
columns need to be input for the function to work. [string]}

\item{lossRateSF6}{Dataframe column name for SF6 loss rate [string]}

\item{peakMaxVelocity}{Dataframe column name for stream velocity calculated with peak 
max method [string]}

\item{meanDepth}{Dataframe column name for mean depth [string]}

\item{meanTemp}{Dataframe column name for mean temperature [string]}

\item{outputSuffix}{Suffix to attach to the names of the reaeration metrics [string]}
}
\value{
This function returns a the input dataframe appended with loss rate, travel time,
SF6 reaeration rate, O2 gas transfer velocity, and Schmidt number 600 for a given eventID.
}
\description{
This function calculates loss rate, travel time, SF6 reaeration rate, O2
gas transfer velocity, and Schmidt number 600.
}
\examples{
#where the data frame "reaFormatted" is already read in
#reaRatesCalc <- def.calc.reaeration(inputFile = reaFormatted,
#dataDir = paste(path.package("reaRate"),"inst\\extdata", sep = "\\"), plot = TRUE)
#where the data is read in from a file in the working directory (also works with a full path)
#reaRatesCalc <- def.calc.reaeration(inputFile =
#system.file("extdata", "reaTestData.csv", package = "reaRate"))
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
def.calc.peakTime for calculating travel times and def.format.reaeration for
formatting reaeration data
}
\author{
Kaelin M. Cawley \email{kcawley@battelleecology.org} \cr
}
\keyword{SF6,}
\keyword{deaeration,}
\keyword{metabolism,}
\keyword{reaeration,}
\keyword{rivers,}
\keyword{streams,}
\keyword{surface}
\keyword{tracer}
\keyword{water,}
