% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.calc.trvl.time.R
\name{def.calc.trvl.time}
\alias{def.calc.trvl.time}
\title{Travel time and mean depth calculations}
\usage{
def.calc.trvl.time(
  inputFile = NULL,
  loggerData = NULL,
  namedLocation = "namedLocation",
  injectionTypeName = "injectionType",
  eventID = "eventID",
  slopeRaw = "slopeRaw",
  slopeClean = "slopeClean",
  slopeSaltCorr = "slopeSaltCorr",
  slopeNormClean = "slopeNormClean",
  slopeNormCorr = "slopeNormCorr",
  stationToInjectionDistance = "stationToInjectionDistance",
  hoboSampleID = "hoboSampleID",
  slugPourTime = "slugPourTime",
  dripStartTime = "dripStartTime",
  meanBackgroundCond = "meanBackgroundCond",
  discharge = "fieldDischarge_lps",
  waterTemp = "waterTemp",
  wettedWidth = "wettedWidth",
  plot = TRUE,
  savePlotPath = NULL
)
}
\arguments{
\item{inputFile}{Name of the data frame containing the information needed to calculate the
reaeration parameters. If the headers are named: "injectionType", "eventID",
"stationToInjectionDistance", "plateauGasConc", "corrPlatSaltConc", "hoboSampleID",
"wettedWidth", respectively, no other inputs are required. Otherwise, the names of the
columns need to be input for the function to work. \link{character}}

\item{loggerData}{User identified filename of logger data \link{character}}

\item{namedLocation}{A string identifier for the station where data was collected \link{character}}

\item{injectionTypeName}{Either constant rate or slug \link{character}}

\item{eventID}{A string identifier to link records collected as part of the same experiment,
SITE.YYYYMMDD for NEON \link{character}}

\item{slopeRaw}{Dataframe column name for SF6 slope from raw data \link{character}}

\item{slopeClean}{Dataframe column name for SF6 slope from data with outliers removed \link{character}}

\item{slopeSaltCorr}{Dataframe column name for SF6 slope from plateau salt corrected data
with outliers removed \link{character}}

\item{slopeNormClean}{Dataframe column name for normalized SF6 slope from data with
outliers removed \link{character}}

\item{slopeNormCorr}{Dataframe column name for normalized SF6 slope from plateau salt
corrected data with outliers removed \link{character}}

\item{stationToInjectionDistance}{Dataframe column name for distance from station to
injection \link{character}}

\item{hoboSampleID}{Dataframe column name for ID to link to conductivity timeseries data \link{character}}

\item{slugPourTime}{Dataframe column name for dateTime when slug was poured \link{character}}

\item{dripStartTime}{Dataframe column name for dateTiem when drip was started \link{character}}

\item{meanBackgroundCond}{Dataframe column name for mean background specific conductance \link{character}}

\item{discharge}{Dataframe column name for stream discharge in literPerSecond \link{character}}

\item{waterTemp}{Dataframe column name for mean water temperature data \link{character}}

\item{wettedWidth}{Dataframe column name for mean wetted width for the stream reach \link{character}}

\item{plot}{User input to plot the SF6/corrected salt concentration versus distance downstream,
defaults to TRUE \link{logical}}

\item{savePlotPath}{If a user specifies a path the plots will be saved to this location \link{character}}
}
\value{
This function returns a list of two dataframes, the input dataframe of data for up to
4 stations per site per date and an output dataframe appended with travel time and mean depth
metrics for a given site and date. See table below for column headers and units for the appended
outputs.\tabular{lll}{
   \strong{columnName} \tab \strong{description} \tab \strong{units} \cr
   S1PeakTime \tab Date and time of conductivity peak at upstream station \tab dateTime \cr
   S4PeakTime \tab Date and time of conductivity peak at downstream station \tab dateTime \cr
   peakMaxTravelTime \tab Difference in upstream and downstream peakTimes \tab second \cr
   btwStaDist \tab Thalweg distance between most upstream and downstream stations \tab meter \cr
   peakMaxVelocity \tab Stream velocity, ratio of btwStaDist and peakMaxTravelTime \tab metersPerSecond \cr
   meanDepth \tab Mean stream depth, meanQ divided by peakMaxVelocity and meanWettedWidth \tab meter \cr
   meanQ_lps \tab Mean stream discharge \tab litersPerSecond \cr
   meanQ_cms \tab Mean stream discharge \tab cubicMetersPerSecond \cr
   meanTemp \tab rmean stream temperature \tab celsius \cr
}
}
\description{
This function calculates travel time and mean depth from reaeration experiments.
}
\examples{
#TBD
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
def.calc.peakTime for calculating travel times and def.format.reaeration for
formatting reaeration data
}
\author{
Kaelin M. Cawley \email{kcawley@battelleecology.org} \cr
}
\keyword{SF6,}
\keyword{deaeration,}
\keyword{metabolism,}
\keyword{reaeration,}
\keyword{rivers,}
\keyword{streams,}
\keyword{surface}
\keyword{tracer}
\keyword{water,}
