% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.format.reaeration.R
\name{def.format.reaeration}
\alias{def.format.reaeration}
\title{Formats reaeration data for rate calculations}
\usage{
def.format.reaeration(
  rea_backgroundFieldCondData,
  rea_backgroundFieldSaltData = NULL,
  rea_fieldData,
  rea_plateauMeasurementFieldData,
  rea_plateauSampleFieldData,
  rea_externalLabDataSalt,
  rea_externalLabDataGas,
  rea_widthFieldData,
  dsc_fieldData = NULL,
  dsc_individualFieldData = NULL,
  dsc_fieldDataADCP = NULL,
  waq_instantaneous = NULL
)
}
\arguments{
\item{rea_backgroundFieldCondData}{This dataframe contains the data for the NEON rea_backgroundFieldCondData table [dataframe]}

\item{rea_backgroundFieldSaltData}{This dataframe contains the data for the NEON rea_backgroundFieldSaltData table [dataframe]}

\item{rea_fieldData}{This dataframe contains the data for the NEON rea_fieldData table [dataframe]}

\item{rea_plateauMeasurementFieldData}{This dataframe contains the data for the NEON rea_plateauMeasurementFieldData table [dataframe]}

\item{rea_plateauSampleFieldData}{This dataframe contains the data for the NEON rea_plateauSampleFieldData table [dataframe]}

\item{rea_externalLabDataSalt}{This dataframe contains the data for the NEON rea_externalLabDataSalt table [dataframe]}

\item{rea_externalLabDataGas}{This dataframe contains the data for the NEON rea_externalLabDataGas table [dataframe]}

\item{rea_widthFieldData}{This dataframe contains the data for the NEON rea_widthFieldData table [dataframe]}

\item{dsc_fieldData}{This dataframe contains the data for the NEON dsc_fieldData table, optional if there is a dsc_fieldDataADCP table [dataframe]}

\item{dsc_individualFieldData}{This dataframe contains the data for the NEON dsc_individualFieldData table, optional [dataframe]}

\item{dsc_fieldDataADCP}{This dataframe contains the data for the NEON dsc_fieldDataADCP table, optional[dataframe]}

\item{waq_instantaneous}{This dataframe contains the data for the NEON Water Quality sensor data, optional[dataframe]}
}
\value{
This function returns one data frame formatted for use with def.calc.reaeration.R
}
\description{
This function formats data from the NEON reaeration data product to calculate
loss rate, travel time, SF6 reaeration rate, O2 gas transfer velocity, and Schmidt number 600.
Either the basic or expanded package can be downloaded. The data files need to be loaded
into the R environment
}
\examples{
#TBD
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
def.calc.tracerTime.R for calculating the stream travel time,
def.plot.reaQcurve.R for plotting reaeration rate versus stream flow
}
\author{
Kaelin M. Cawley \email{kcawley@battelleecology.org} \cr
}
\keyword{gas}
\keyword{number}
\keyword{reaeration,}
\keyword{rivers,}
\keyword{schmidt}
\keyword{streams,}
\keyword{surface}
\keyword{transfer}
\keyword{velocity,}
\keyword{water,}
