% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gas.loss.rate.plot.R
\name{gas.loss.rate.plot}
\alias{gas.loss.rate.plot}
\title{Loss rate plots for QAQC}
\usage{
gas.loss.rate.plot(
  inputFile = NULL,
  injectionTypeName = "injectionType",
  eventID = "eventID",
  stationToInjectionDistance = "stationToInjectionDistance",
  plateauGasConc = "plateauGasConc",
  corrPlatSaltConc = "corrPlatSaltConc",
  savePlotPath = NULL
)
}
\arguments{
\item{inputFile}{Name of the data frame containing the information needed to calculate the
reaeration parameters. If the headers are named: "injectionType", "eventID",
"stationToInjectionDistance", "plateauGasConc", "corrPlatSaltConc", "hoboSampleID",
"wettedWidth", respectively, no other inputs are required. Otherwise, the names of the
columns need to be input for the function to work. [string]}

\item{injectionTypeName}{Either constant rate or slug [string]}

\item{eventID}{A string identifier to link records collected as part of the same experiment,
SITE.YYYYMMDD for NEON [string]}

\item{stationToInjectionDistance}{Dataframe column name for distance from station to
injection [string]}

\item{plateauGasConc}{Dataframe column name for natural log of gas concentration normalized to
background corrected salt concentration [string]}

\item{corrPlatSaltConc}{Dataframe column name for natural log of gas concentration normalized to
background corrected salt concentration [string]}

\item{savePlotPath}{If a user specifies a path the plots will be saved to this location [string]}
}
\value{
This function returns a list of two dataframes, the input dataframe of data for up to
4 stations per site per date and an output dataframe appended with loss rate, travel time,
SF6 reaeration rate, O2 gas transfer velocity, and Schmidt number 600 for a given site and date
}
\description{
This function calculates loss rate, travel time, SF6 reaeration rate, O2
gas transfer velocity, and Schmidt number 600.
}
\examples{
#where the data frame "reaFormatted" is already read in
#reaRatesCalc <- def.calc.reaeration(inputFile = reaFormatted,
#dataDir = paste(path.package("reaRate"),"inst\\extdata", sep = "\\"), plot = TRUE)
#where the data is read in from a file in the working directory (also works with a full path)
#reaRatesCalc <- def.calc.reaeration(inputFile =
#system.file("extdata", "reaTestData.csv", package = "reaRate"))
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
def.calc.peakTime for calculating travel times and def.format.reaeration for
formatting reaeration data
}
\author{
Kaelin M. Cawley \email{kcawley@battelleecology.org} \cr
}
\keyword{SF6,}
\keyword{deaeration,}
\keyword{metabolism,}
\keyword{reaeration,}
\keyword{rivers,}
\keyword{streams,}
\keyword{surface}
\keyword{tracer}
\keyword{water,}
