% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slugTime.R
\name{slugTime}
\alias{slugTime}
\title{Slug peak concentration location}
\usage{
slugTime(condData, pkRange = 5, pkNum = 4)
}
\arguments{
\item{condData}{data to be fit for peak}

\item{pkRange}{Number of points on either side of peak to test for slope  [integer]}

\item{pkNum}{Number of points that must be increasing/decreasing for peak slope [integer]}
}
\value{
This function returns the peak tracer index [integer]
}
\description{
This function determines the location of the peak specific conductance for use 
calculating travel time between an upstream and downstream sensor set.
}
\examples{
#Using an example file
s4peakLoc <- slugTime(condDataS4)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
def.calc.reaeration.R for calculating reaeration rates and 
def.calc.travelTime.R for determining travel times
}
\author{
Kaelin M. Cawley \email{kcawley@battelleecology.org} \cr
}
\keyword{metabolism}
\keyword{reaeration,}
\keyword{rivers,}
\keyword{streams,}
\keyword{surface}
\keyword{time,}
\keyword{travel}
\keyword{velocity,}
\keyword{water,}
