% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkgd.salt.conc.plot.R
\name{bkgd.salt.conc.plot}
\alias{bkgd.salt.conc.plot}
\title{Background Salt Plot for QAQC Decisions}
\usage{
bkgd.salt.conc.plot(inputFile = NULL, savePlotPath = NULL)
}
\arguments{
\item{inputFile}{Name of the data frame containing the information needed to calculate the
reaeration parameters. If the headers are named: "injectionType", "eventID",
"stationToInjectionDistance", "plateauGasConc", "corrPlatSaltConc", "hoboSampleID",
"wettedWidth", respectively, no other inputs are required. Otherwise, the names of the
columns need to be input for the function to work. [string]}

\item{savePlotPath}{If a user specifies a path the plots will be saved to this location [string]}
}
\value{
This function returns a list of two dataframes, the input dataframe of data for up to
4 stations per site per date and an output dataframe appended with loss rate, travel time,
SF6 reaeration rate, O2 gas transfer velocity, and Schmidt number 600 for a given site and date
}
\description{
This function calculates loss rate, travel time, SF6 reaeration rate, O2
gas transfer velocity, and Schmidt number 600.
}
\examples{
#TBD
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
def.calc.peakTime for calculating travel times and def.format.reaeration for
formatting reaeration data
}
\author{
Kaelin M. Cawley \email{kcawley@battelleecology.org} \cr
}
\keyword{SF6,}
\keyword{deaeration,}
\keyword{metabolism,}
\keyword{reaeration,}
\keyword{rivers,}
\keyword{streams,}
\keyword{surface}
\keyword{tracer}
\keyword{water,}
