% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.calc.trvl.time.R
\name{def.calc.trvl.time}
\alias{def.calc.trvl.time}
\title{Travel time and mean depth calculations}
\usage{
def.calc.trvl.time(
  inputFile = NULL,
  loggerData = NULL,
  namedLocation = "namedLocation",
  injectionTypeName = "injectionType",
  eventID = "eventID",
  slopeRaw = "slopeRaw",
  slopeClean = "slopeClean",
  slopeSaltCorr = "slopeSaltCorr",
  slopeNormClean = "slopeNormClean",
  slopeNormCorr = "slopeNormCorr",
  stationToInjectionDistance = "stationToInjectionDistance",
  hoboSampleID = "hoboSampleID",
  slugPourTime = "slugPourTime",
  dripStartTime = "dripStartTime",
  meanBackgroundCond = "meanBackgroundCond",
  discharge = "fieldDischarge_lps",
  waterTemp = "waterTemp",
  wettedWidth = "wettedWidth",
  plot = TRUE,
  savePlotPath = NULL
)
}
\arguments{
\item{inputFile}{Name of the data frame containing the information needed to calculate the
reaeration parameters. If the headers are named: "injectionType", "eventID",
"stationToInjectionDistance", "plateauGasConc", "corrPlatSaltConc", "hoboSampleID",
"wettedWidth", respectively, no other inputs are required. Otherwise, the names of the
columns need to be input for the function to work. [string]}

\item{loggerData}{User identified filename of logger data [string]}

\item{namedLocation}{A string identifier for the station where data was collected [string]}

\item{injectionTypeName}{Either constant rate or slug [string]}

\item{eventID}{A string identifier to link records collected as part of the same experiment,
SITE.YYYYMMDD for NEON [string]}

\item{slopeRaw}{Dataframe column name for SF6 slope from raw data [string]}

\item{slopeClean}{Dataframe column name for SF6 slope from data with outliers removed [string]}

\item{slopeSaltCorr}{Dataframe column name for SF6 slope from plateau salt corrected data 
with outliers removed [string]}

\item{slopeNormClean}{Dataframe column name for normalized SF6 slope from data with 
outliers removed [string]}

\item{slopeNormCorr}{Dataframe column name for normalized SF6 slope from plateau salt 
corrected data with outliers removed [string]}

\item{stationToInjectionDistance}{Dataframe column name for distance from station to
injection [string]}

\item{hoboSampleID}{Dataframe column name for ID to link to conductivity timeseries data [string]}

\item{slugPourTime}{Dataframe column name for dateTime when slug was poured [string]}

\item{dripStartTime}{Dataframe column name for dateTiem when drip was started [string]}

\item{meanBackgroundCond}{Dataframe column name for mean background specific conductance [string]}

\item{discharge}{Dataframe column name for stream discharge in literPerSecond [string]}

\item{waterTemp}{Dataframe column name for mean water temperature data [string]}

\item{wettedWidth}{Dataframe column name for mean wetted width for the stream reach [string]}

\item{plot}{User input to plot the SF6/corrected salt concentration versus distance downstream,
defaults to TRUE [boolean]}

\item{savePlotPath}{If a user specifies a path the plots will be saved to this location [string]}
}
\value{
This function returns a list of two dataframes, the input dataframe of data for up to
4 stations per site per date and an output dataframe appended with travel time and mean depth
for a given site and date
}
\description{
This function calculates travel time and mean depth from reaeration experiments.
}
\examples{
#TBD
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
def.calc.peakTime for calculating travel times and def.format.reaeration for
formatting reaeration data
}
\author{
Kaelin M. Cawley \email{kcawley@battelleecology.org} \cr
}
\keyword{SF6,}
\keyword{deaeration,}
\keyword{metabolism,}
\keyword{reaeration,}
\keyword{rivers,}
\keyword{streams,}
\keyword{surface}
\keyword{tracer}
\keyword{water,}
