% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.data.resolveDupes.R
\name{def.data.resolveDupes}
\alias{def.data.resolveDupes}
\title{Resolves duplicate lab analysis results - multiple results for the same sample name}
\usage{
def.data.resolveDupes(filepath = "", tableName = "")
}
\arguments{
\item{filepath}{The path to the directory containing the lab data to de-duplicate [string]}

\item{tableName}{The name of the table in the file path directory to de-duplicate [string]}
}
\value{
This function returns one data frame formatted for use with def.calc.reaeration.R
}
\description{
This includes multiple user-selectable options for resolving duplicate lab
samples: 1) keeping the most recent measurement for each duplicate sample.
}
\examples{
def.data.resolveDupes(filepath = "GitHub/NEON-reaeration/filesToStack20190/stackedFiles", tableName = "externalLabDataSalt")
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\author{
Geoffrey L. House \email{housegl@battelleecology.org} \cr
}
\keyword{gas}
\keyword{number}
\keyword{reaeration,}
\keyword{rivers,}
\keyword{schmidt}
\keyword{streams,}
\keyword{surface}
\keyword{transfer}
\keyword{velocity,}
\keyword{water,}
